requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
asgiref==3.8.1
attrs==25.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.7
execnet==2.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.2
packaging==24.1
pandas==2.2.3
playwright==1.51.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.3
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-json-ctrf==0.3.5
pytest-metadata==3.1.1
pytest-playwright==0.7.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
